pam_oath: Don't require pam_oath.so to be installed for root tests

+++ pam_oath/tests/run-root-tests.sh
@@ -37,7 +37,18 @@ if test -f $ETCUSRCFG; then
     exit 1
 fi
 
-echo "auth requisite pam_oath.so debug usersfile=$ETCUSRCFG window=20 digits=6" > $ETCPAMCFG
+so_path_rel="${srcdir}/../.libs/pam_oath.so"
+so_path="$(readlink -f "${so_path_rel}")"
+if test -z "${so_path}"; then
+    echo "Unable to resolve path to pam_oath.so: ${so_path_rel}"
+    exit 1
+fi
+if ! test -f "${so_path}"; then
+    echo "pam_oath.so not found at: ${so_path}"
+    exit 1
+fi
+
+echo "auth requisite [${so_path}] debug usersfile=$ETCUSRCFG window=20 digits=6" > $ETCPAMCFG
 echo "HOTP user1 - 00" > $ETCUSRCFG
 echo "HOTP user2 pw 00" >> $ETCUSRCFG
 echo "HOTP/T30 user3 - 00" >> $ETCUSRCFG
